removed jump bind, resolved build warnings, refactored settings menu to separate file  (#169)
## Description
```
Removed jump text and bind from settings menu.
Cleaned up basic build warnings.
Created and refactored all code related to settings menu to settings_menu.cpp from game.cpp
Removed all jump functionality.
Fixed char select up/down to W/S and UP/DOWN permanently. 
```
Issue: #182 

## Checklist 
```
Both the requestee and the reviewer should review this checklist
```
- [x] Code Compiles on your machine
- [x] PR Filled in (remove the code formatted blocks)
- [x] Code is appropriately compartmentalized with 
- [x] feature has no/minimal interference for other functions
- [x] Code is well documented with comments
- [x] Code is readable
- [ ] Tests (if included) pass

## Additional Notes
```
Add any additional notes or context that may be helpful for reviewers.
```

---------

+++ src/renderer.cpp
@@ -462,27 +462,30 @@ void GlRender::handleTexturedRenders()
 
 void GlRender::handleNotifications(float elapsed_ms)
 {
-    for (Entity& notification_entity : registry.notifications.entities) {
-        Notification& notification = registry.notifications.get(notification_entity);
+    for (Entity &notification_entity : registry.notifications.entities)
+    {
+        Notification &notification = registry.notifications.get(notification_entity);
 
         float opacity = 1.0f;
         notification.counter_ms -= elapsed_ms;
 
-        if (notification.counter_ms <= 0.f) {
+        if (notification.counter_ms <= 0.f)
+        {
             registry.remove_all_components_of(notification_entity);
             continue;
         }
 
-        if (notification.counter_ms < 150.f) {
+        if (notification.counter_ms < 150.f)
+        {
             opacity = notification.counter_ms / 150.f;
         }
 
-        float x = notification.player1Side ? 0 : M_WINDOW_WIDTH_PX / 2;
+        int x = notification.player1Side ? 0 : M_WINDOW_WIDTH_PX / 2;
         float y = 15.f * opacity;
 
         renderTexturedQuadScaled(
             notification.texture_id,
-            x, y,
+            static_cast<float>(x), y,
             M_WINDOW_WIDTH_PX / 2, M_WINDOW_HEIGHT_PX,
             1.0f, opacity);
     }
@@ -693,7 +696,7 @@ void GlRender::renderUI(int timer)
         1.0f);
 
     // Render timer background
-    //renderTexturedQuadScaled(
+    // renderTexturedQuadScaled(
     //    m_timerBackgroundTexture,
     //    timerBgX, timerBgY,
     //    timerBgWidth, timerBgHeight,
@@ -729,8 +732,8 @@ void GlRender::renderUI(int timer)
     handleP1Health(p1Health);
     handleP2Health(p2Health);
 
-    handleP1Posture(p1Posture);
-    handleP2Posture(p2Posture);
+    handleP1Posture(static_cast<int>(p1Posture));
+    handleP2Posture(static_cast<int>(p2Posture));
 
     // render timer
     float textWidth = std::to_string(timer).length() * 30.0f * xscale;
